Restore critical services and clean up unused code
Restores ProseMirror rendering and version checking as dedicated services, following architecture principles. Removes unused and legacy utility files, updates imports across components, and documents the changes in architecture and refactor tracking files. Ensures mandatory rendering and version checking functionality is preserved for Atlassian content and extension updates.
