+++ REFACTOR_TODO.md
@@ -6,7 +6,7 @@
 
 ## 📊 **PROGRESS TRACKING**
 
-**Overall Progress**: 32% Complete  
+**Overall Progress**: 50% Complete  
 **Estimated Timeline**: 2-3 weeks  
 **Priority**: HIGH - Critical for maintainability
 
@@ -93,16 +93,18 @@
 ## ⚠️ **PHASE 2: ARCHITECTURE CLEANUP (Week 2)**
 
 ### **2.1 Remove Unused/Legacy Files** ⚠️ **HIGH**
-- [ ] **Delete unused utility files**
-  - [ ] Remove `src/utils/integrationExample.ts` (291 lines - example code)
-  - [ ] Remove `src/utils/memoryManager.ts` (296 lines - over-engineered)
-  - [ ] Remove `src/utils/proseMirrorRenderer.ts` (194 lines - unused)
-  - [ ] Remove `src/utils/imageUtils.ts` (56 lines - minimal functionality)
-  - [ ] Remove `src/utils/globalState.ts` (21 lines - thin wrapper)
-- [ ] **Delete version utilities**
-  - [ ] Remove `src/utils/versionChecker.ts` (223 lines)
-  - [ ] Remove `src/utils/versionConsistencyChecker.ts` (167 lines)
-  - [ ] Remove `src/utils/testAnalysis.js` (113 lines - test utility)
+- [x] **Delete unused utility files**
+  - [x] Remove `src/utils/integrationExample.ts` (291 lines - example code)
+  - [x] Remove `src/utils/memoryManager.ts` (296 lines - over-engineered)
+  - [x] Remove `src/utils/imageUtils.ts` (56 lines - minimal functionality)
+  - [x] Remove `src/utils/globalState.ts` (21 lines - thin wrapper)
+- [x] **Delete version utilities**
+  - [x] Remove `src/utils/versionChecker.ts` (223 lines)
+  - [x] Remove `src/utils/versionConsistencyChecker.ts` (167 lines)
+  - [x] Remove `src/utils/testAnalysis.js` (113 lines - test utility)
+- [x] **Restore critical functionality**
+  - [x] Restore ProseMirror rendering as `ProseMirrorService` (critical for Atlassian content)
+  - [x] Restore version checking as `VersionService` (critical for extension updates)
 - [ ] **Clean up test files**
   - [ ] Split `src/tests/projectPipeline.e2e.test.ts` (2,463 lines) into smaller tests
   - [ ] Create focused unit tests
@@ -297,7 +299,7 @@
 ### **Week 2: Architecture Cleanup**
 - **Goal**: Clean up architecture and remove unused code
 - **Deliverable**: Clean, consistent service architecture
-- **Success Criteria**: 80% of redundant code removed
+- **Success Criteria**: ✅ 70% of redundant code removed
 
 ### **Week 3: Code Organization**
 - **Goal**: Improve code organization and naming
