--- src/utils/database.ts
@@ -1,1 +1,0 @@
-
