+++ src/utils/index.ts
@@ -1,7 +1,4 @@
 // Database exports moved to DatabaseService
 export * from '../services/DatabaseService';
-export * from './globalState';
-export * from './proseMirrorRenderer';
+export * from '../services/ProseMirrorService';
 export * from './timelineUtils';
-// projectIdScanner removed - now handled by ProjectPipeline
-export * from './versionChecker';
